refactor: move tag to tags file
