+++ src/islands/Checkbox.tsx
@@ -2,7 +2,7 @@ import { IS_BROWSER } from "$fresh/runtime.ts";
 import { Field, Input, Label } from "@headlessui/react";
 import { clsx } from "clsx";
 import type { JSX } from "preact";
-import { tw } from "../utils/tailwind.ts";
+import { tw } from "../utils/tags.ts";
 
 export interface CheckboxProps {
   readonly name: string;
