+++ src/islands/ScrollDown.tsx
@@ -4,7 +4,7 @@ import { clsx } from "clsx";
 import type { JSX } from "preact";
 import { IconChevronDown } from "../components/icons.ts";
 import { floatingButtonStyles } from "../components/styles.ts";
-import { tw } from "../utils/tailwind.ts";
+import { tw } from "../utils/tags.ts";
 
 function scrollDown(): void {
   globalThis.scrollTo({ top: globalThis.innerHeight });
